requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ML Frameworks
xgboost>=2.0.0
lightgbm>=4.0.0

# MLOps Tools
mlflow>=2.8.0
wandb>=0.15.0
dvc>=3.0.0
dvc-s3>=2.22.0  # For S3 remote storage (optional)

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Data Validation
great-expectations>=0.18.0

# Hyperparameter Tuning
optuna>=3.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.26.0

# Monitoring & Logging
prometheus-client>=0.19.0
